refactor(connlib): share authorization message bookkeeping
The Authorization init payload, the group-by-client fold consuming it,
and the past-expiry-collapses-to-zero-TTL logic each exist once per
role. Client-to-client resync (#14060) needs all three on the client
too, so move them to shared code instead of copying them.
